requirements.txt
Flask==2.3.3
scikit-learn==1.3.0
Pillow
numpy==1.24.3
